+++ client/Navigation/ControlPanelNavigation.tsx
@@ -2,8 +2,8 @@ import { Navbar, ScrollArea } from '@mantine/core'
 import { useViewportSize } from '@mantine/hooks'
 import { FaQuestion } from '@react-icons/all-files/fa/FaQuestion'
 import React from 'react'
+import { DatabaseFolder } from '../../shared/database'
 import { ButtonIcon } from '../Components/ControlPanel/ButtonIcon'
-import { DatabaseFolder } from '../../database/repository/folder'
 import { useAppSelector } from '../apis/redux/mainStore'
 import { NavbarItem } from './NavbarItem'
 import { useStyles } from './Styles'
