+++ client/Pages/Admin/Presets.tsx
@@ -37,7 +37,6 @@ import { FaTrash } from '@react-icons/all-files/fa/FaTrash'
 import React, { useEffect, useState } from 'react'
 import { DragDropContext, Draggable, Droppable } from 'react-beautiful-dnd'
 import { usePrompt } from '../../apis/utilities/usePrompt'
-import { DatabasePreset, PresetTypes } from '../../../database/repository/preset'
 import { E131PresetEditModal } from '../../Components/Admin/Controls/Presets/EditModal/E131'
 import { HTTPPresetEditModal } from '../../Components/Admin/Controls/Presets/EditModal/HTTP'
 import { MacroPresetEditModal } from '../../Components/Admin/Controls/Presets/EditModal/Macro'
@@ -46,6 +45,7 @@ import { isValidJson } from '../../Components/Admin/Controls/Presets/EditModal/i
 import { ButtonIconSelectItem, availableIcons } from '../../Components/ControlPanel/ButtonIcon'
 import { useAppSelector } from '../../apis/redux/mainStore'
 import { ApiCall } from '../../apis/wrapper'
+import { DatabasePreset, PresetTypes } from '../../../shared/database'
 
 interface FormValues {
 	presets: Array<DatabasePreset>
