+++ client/Pages/ControlPanel/Preset.tsx
@@ -2,8 +2,8 @@ import { Button, Paper } from '@mantine/core'
 import { FaLevelUpAlt } from '@react-icons/all-files/fa/FaLevelUpAlt'
 import React from 'react'
 import { Link, useNavigate, useParams } from 'react-router-dom'
+import { DatabaseFolder } from '../../../shared/database'
 import { ButtonIcon } from '../../Components/ControlPanel/ButtonIcon'
-import { DatabaseFolder } from '../../../database/repository/folder'
 import { PresetFaders } from '../../Components/ControlPanel/PresetFaders'
 import { DangerouslySetHTML } from '../../Components/DangerouslySetHTML'
 import { useAppSelector } from '../../apis/redux/mainStore'
