+++ client/apis/redux/databaseSlice.ts
@@ -1,9 +1,9 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit'
-import { Database } from '../../../api/database'
+import { Database } from '../../../shared/database'
 import { ApiCall } from '../wrapper'
 import store from './mainStore'
 
-const initialState: Database | null = null
+const initialState: Database | null = null as Database | null
 const getFromAPIFatActionCreator = (): void => {
 	// https://redux.js.org/faq/code-structure/#how-should-i-split-my-logic-between-reducers-and-action-creators-where-should-my-business-logic-go
 	ApiCall.get('/database', {}).then(response => {
