+++ client/package.json
@@ -1,5 +1,5 @@
 {
-	"name": "paradisepi-server",
+	"name": "paradisepi-client",
 	"private": false,
 	"version": "3.0.0",
 	"type": "module",
@@ -34,7 +34,7 @@
 		"@mantine/rte": "~5.10.5",
 		"@react-icons/all-files": "^4.1.0",
 		"@reduxjs/toolkit": "^1.9.5",
-    "axios": "^1.6.0",
+		"axios": "^1.6.0",
 		"qrcode.react": "^3.1.0",
 		"react": "^18.2.0",
 		"react-beautiful-dnd": "^13.1.1",
@@ -47,10 +47,10 @@
 	"devDependencies": {
 		"@types/react": "^18.2.8",
 		"@types/react-dom": "^18.2.4",
-		 "@types/react-beautiful-dnd": "^13.1.4",
+		"@types/react-beautiful-dnd": "^13.1.4",
 		"@typescript-eslint/eslint-plugin": "^5.59.8",
 		"@typescript-eslint/parser": "^5.59.8",
-    "@vitejs/plugin-react": "^4.3.1",
+		"@vitejs/plugin-react": "^4.3.1",
 		"eslint": "^8.42.0",
 		"eslint-config-google": "^0.14.0",
 		"eslint-config-prettier": "^8.8.0",
@@ -63,7 +63,7 @@
 		"postcss-preset-mantine": "1.15.0",
 		"postcss-simple-vars": "^7.0.1",
 		"vite": "^5.3.1",
-    "css-loader": "^6.8.1",
+		"css-loader": "^6.8.1",
 		"postcss": "^8.4.31",
 		"postcss-import": "^15.1.0",
 		"postcss-loader": "^7.3.2",
