+++ server/package.json
@@ -3,13 +3,13 @@
 	"private": false,
 	"version": "3.0.0",
 	"description": "Raspberry Pi based sACN lighting/OSC sound facility control panel",
-	"main": "dist/index.js",
-	"types": "dist/index.d.ts",
+	"main": "dist/server/src/index.js",
+	"types": "dist/server/src/index.d.ts",
 	"scripts": {
 		"prepublish": "npm run build",
 		"build": "tsc",
 		"test": "echo \"Error: no test specified\" && exit 1",
-		"develop": "ts-node src/index.ts",
+		"develop": "tsx src/index.ts",
 		"lint": "eslint --ext .ts,.tsx .",
 		"typeorm": "typeorm-ts-node-commonjs -d src/database/dataSource.ts"
 	},
@@ -31,6 +31,7 @@
 		"@paradise-pi/e131": "~1.3.0",
 		"axios": "^1.6.0",
 		"better-sqlite3": "^8.4.0",
+		"formidable": "^3.2.5",
 		"ip": "^1.1.9",
 		"node-os-utils": "^1.3.7",
 		"node-static": "^0.7.11",
@@ -38,14 +39,14 @@
 		"portfinder": "^1.0.32",
 		"reflect-metadata": "^0.1.14",
 		"socket.io": "^4.6.2",
+		"ts-node": "^10.9.2",
 		"typeorm": "^0.3.20",
-		"winston": "^3.9.0",
-		"formidable": "^3.2.5"
+		"winston": "^3.9.0"
 	},
 	"devDependencies": {
+		"@types/formidable": "^2.0.6",
 		"@types/ip": "^1.1.0",
 		"@types/node": "^18.19.39",
-		"@types/formidable": "^2.0.6",
 		"@types/node-os-utils": "^1.3.1",
 		"@types/node-static": "^0.7.7",
 		"@typescript-eslint/eslint-plugin": "^5.59.8",
@@ -64,7 +65,7 @@
 		"prettier": "^2.8.8",
 		"style-loader": "^3.3.3",
 		"ts-loader": "^9.4.3",
-		"ts-node": "^10.9.1",
+		"tsx": "^4.15.7",
 		"typescript": "^4.9.5"
 	}
 }
