+++ server/src/api/fader/faderRouter.ts
@@ -1,7 +1,8 @@
+import { Database } from '../../../../shared/database'
 import { DatabaseFader } from '../../../../shared/sharedTypes'
 import { FaderRepository } from '../../database/repository/fader'
 import logger from '../../logger'
-import { createDatabaseObject, Database, sendDatabaseObject } from '../database'
+import { createDatabaseObject, sendDatabaseObject } from '../database'
 /**
  * This is a REST router for the fader API.
  * @param path - The path requested by the original route requestor
