+++ server/src/webServer/index.html
@@ -1,0 +1,1 @@
+Test - need to figure out at some point how to serve the files from the client folder.
