+++ shared/meterFunctions.ts
